requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
drf-yasg==1.21.4
environs==8.0.0
frozenlist==1.3.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
multidict==6.0.2
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-decouple==3.7
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
types-cryptography==3.3.23.1
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
